+++ CHANGELOG.md
@@ -54,6 +54,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - entity data objects have only an empty constructor now
     - creation of entity data objects moved to factories to allow extensibility
 
+### [shopsys/framework]
+#### Fixed
+- [#291 - Unnecessary SQL queries on category detail in admin](https://github.com/shopsys/shopsys/pull/304):
+    - category translations for ancestor category are loaded in the same query as categories
+
 ## [7.0.0-alpha3] - 2018-07-03
 ### [shopsys/framework]
 #### Changed